requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gitdb2==4.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyYAML==5.3
requests==2.22.0
smmap2==2.0.5
urllib3==1.25.7
Werkzeug==1.0.0
